forge-ai/requirements.txt
torch==2.1.0
torchvision==0.16.0
diffusers==0.25.0
transformers==4.36.0
accelerate==0.25.0
safetensors==0.4.1
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.26.2
pandas==2.1.3
requests==2.31.0
tqdm==4.66.1
pyyaml==6.0.1
selenium==4.15.2
beautifulsoup4==4.12.2
webdriver-manager==4.0.1
